page.sveltes webinar pages fully components
♻️   refactor: all sections of the webinars and his slug page are now in their own component page
