+++ src/lib/Bricks/webinar-header.svelte
@@ -1,0 +1,8 @@
+<h1>Webinars</h1>
+
+<style>
+    h1 {
+    font-size: var(--font-size-sm);
+    opacity: 0;
+  }
+</style>
